refactor: move compression test to response module and remove unused getters
- Removed public headers() and body() getters from Response
- Made Response fields pub(crate) for internal access within the crate
- Moved test_gzip_compression to src/response/tests.rs
- Added const len() method to HttpBody
