+++ src/response/tests.rs
@@ -346,3 +346,36 @@ fn test_set_encoding_overwrites_existing() {
     assert!(output.to_lowercase().contains("content-encoding: gzip"));
     assert!(!output.to_lowercase().contains("content-encoding: deflate"));
 }
+
+#[test]
+fn test_gzip_compression() {
+    use crate::server::compress_response_body;
+    use flate2::read::GzDecoder;
+    use std::io::Read;
+
+    let mut resp = Response::new(HttpStatus::Ok);
+    let original_body = "Hello, world! This is a test string to be compressed.";
+    resp.set_str_body(original_body);
+
+    // Call the compression function
+    compress_response_body(&mut resp).unwrap();
+
+    // Verify headers (accessing pub(crate) field directly)
+    assert_eq!(resp.headers.get("Content-Encoding"), Some("gzip"));
+
+    // Extract body and decompress it
+    if let HttpBody::Content(compressed_bytes) = &resp.body {
+        let mut decoder = GzDecoder::new(&compressed_bytes[..]);
+        let mut decompressed_body = String::new();
+        decoder.read_to_string(&mut decompressed_body).unwrap();
+        assert_eq!(decompressed_body, original_body);
+
+        // Verify Content-Length is updated to compressed size
+        assert_eq!(
+            resp.headers.content_length().unwrap(),
+            compressed_bytes.len()
+        );
+    } else {
+        panic!("Response body should be Content");
+    }
+}
